Soften light mode + make default dark mode
Light mode was a bit harsh, so soften text and make the background a warmer white. (It's still an awful lot of white on the home page, which is stark.) Switch things around so tht dark mode is the default and light mode needs to be set by the user agent.